refactoring - ServiceWrapper in separate file
