+++ requestor/service_manager/__init__.py
@@ -1,1 +1,2 @@
 from .yagna_erigon_manager import YagnaErigonManager
+from .service_wrapper import ServiceWrapper
