+++ requestor/service_manager/services/__init__.py
@@ -1,2 +1,2 @@
-from .erigon import Erigon
-from .pseudo_erigon import PseudoErigon
+from .erigon import ErigonService
+# from .pseudo_erigon import PseudoErigon
