+++ requestor/service_manager/services/erigon.py
@@ -1,25 +1,8 @@
-import asyncio
-from datetime import datetime
-from dataclasses import dataclass, field
-from typing import Optional, Dict
 import json
 
 from .erigon_payload import ErigonPayload
 
 from yapapi.executor.services import Service
-
-SECONDS_BETWEEN_UPDATES = 1
-
-
-@dataclass
-class RuntimeState():
-    status: str
-    url: Optional[str] = None
-    auth: Optional[Dict[str, str]] = None
-    timestamp: datetime = field(init=False)
-
-    def __post_init__(self):
-        self.timestamp = datetime.now()
 
 
 class ErigonService(Service):
@@ -60,72 +43,3 @@ class ErigonService(Service):
         mock_echo_data, erigon_data = stdout.split('ERIGON: ', 2)
         erigon_data = json.loads(erigon_data)
         return erigon_data
-
-
-class Erigon():
-    service_cls = ErigonService
-
-    def __init__(self):
-        self.stopped = False
-        self.runtime_state = RuntimeState('initializing')
-        self.update_task = asyncio.create_task(self.update_state())
-        self.service = None
-
-    def set_service(self, service):
-        self.service = service
-
-    @property
-    def started(self):
-        return self.service is not None
-
-    @property
-    def id(self):
-        if self.service:
-            return self.service.id
-        return '[NOT STARTED YET]'
-
-    async def status(self):
-        return await self.run_single_command('STATUS')
-
-    async def stop(self):
-        if self.stopped:
-            return
-
-        self.disable()
-        if self.started:
-            return await self.run_single_command('STOP')
-        else:
-            return "not started yet"
-
-    def disable(self):
-        if self.stopped:
-            return
-        self.stopped = True
-        print(f"STOPPING {self}")
-
-    async def run_single_command(self, cmd):
-        self.service.send_message_nowait(cmd)
-        service_signal = await self.service.receive_message()
-        #   TODO: how do we check if we got response for the signal sent?
-        #         this is irrelevant now, but could be useful in the future
-        return service_signal.message
-
-    async def update_state(self):
-        while True:
-            if self.stopped:
-                self.runtime_state = RuntimeState('stopping')
-                break
-            if not self.started:
-                self.runtime_state = RuntimeState('starting')
-                await asyncio.sleep(1)
-                continue
-
-            res = await self.status()
-            self.runtime_state = RuntimeState(**res)
-            if not self.stopped:
-                #   Additional check for self.stopped because this could have changed
-                #   while we awaited for self.status()
-                await asyncio.sleep(SECONDS_BETWEEN_UPDATES)
-
-    def __repr__(self):
-        return f"{type(self).__name__}[id={self.id}]"
